// src/app/state/booking/booking.action.types.ts
export enum BookingActionTypes {
    BookingSuccessNotification = '[Booking] Success Notification',
    BookingFailureNotification = '[Booking] Failure Notification',
    CreateReservation= '[Reservation] Create Reservation',
    CreateReservationSuccess = '[Reservation] Create Reservation Success',
    CreateReservationFailure = '[Reservation] Create Reservation Failure',
    AddBookingToStore ='[Booking] Add Booking To Store',
    CreateBooking = '[Booking] Create Booking',
    CreateBookingSuccess = '[Booking] Create Booking Success',
    CreateBookingFailure = '[Booking] Create Booking Failure',
    GetCustomerBookings = '[Booking] Get Customer Bookings',
    GetCustomerBookingsSuccess = '[Booking] Get Customer Bookings Success',
    GetCustomerBookingsFailure = '[Booking] Get Customer Bookings Failure',
    GetMerchantBookings = '[Booking] Get Merchant Bookings',
    GetMerchantBookingsSuccess = '[Booking] Get Merchant Bookings Success',
    GetMerchantBookingsFailure = '[Booking] Get Merchant Bookings Failure',
    GetBookingTimes = '[Booking] Get Booking Times',
    GetBookingTimesSuccess = '[Booking] Get Booking Times Success',
    GetBookingTimesFailure = '[Booking] Get Booking Times Failure',
};